Dev (#8)
* Node.js Basic structure

* Node.js add basic structure Router for notes/ Models for notes

* changed .gitignore

* Dev starodubets (#4)

* Node.js add basic structure Router for notes/ Models for notes

* Node.js add basic structure Router for notes/ Models for notes

* Add method
 get create: /notes-create,
 post add to Mongo

* Add method get create: /notes-create/:id to change notes

* Add method delete : /notes,

* Add basic method get, post : /image,

* Add method delete : /image, add picture in notes

* Add method put  : /

* Task 3 (#5)

* added form for todo list

* added frontend for todo lists and post request

* Dev starodubets (#6)

* Node.js add basic structure Router for notes/ Models for notes

* Node.js add basic structure Router for notes/ Models for notes

* Add method
 get create: /notes-create,
 post add to Mongo

* Add method get create: /notes-create/:id to change notes

* Add method delete : /notes,

* Add basic method get, post : /image,

* Add method delete : /image, add picture in notes

* Add method put  : /

* Add all method for image

* Add all method for image

* Add layout form index.html, datailsNotes, createNotes

* Fix bugs

* fixed syntax errors

* notes fix

* notes fixes

* fixed index

* fixed images conditions in detailsNote.pug and layout.pug

* added styles

* renamed classes, files, variables

* added styles for label

* Final part Notes

* Final part List/ contains a problem put/ (#7)

* .gitignore changed

* added yarn.lock to .gitignore

* changed todoList.js, lists.pug

* changed inputs behavior

* fixed todo lists

* removed redundant commented code

* refactored code of todo lists

* fixed carriage behavior after entering Backspase, added onInput event

+++ public/js/main.js
@@ -1,53 +1,62 @@
 let form = document.getElementById('addToMongo');
 
 if (form) {
-    form.addEventListener('submit', function (event) {
+    form.addEventListener('submit', onFormSubmit);
 
-            event.preventDefault();
 
-            let data = {};
-            let formData = new FormData(this);
+    function onFormSubmit(event) {
+        let form = document.getElementById('addToMongo');
+        event.preventDefault();
 
-            for (let [name, value] of formData) {
+        let data = {};
+        let formData = new FormData(form);
 
-               if(typeof value === "object"){
-                   formData.append('nameImg', value.name)
-               }
-               data[name] = value;
+        for (let [name, value] of formData) {
+
+            if (typeof value === "object") {
+                formData.append('nameImg', value.name);
+
+                if (value.name.length === 0 && data.id) {
+                    let file = document.getElementById('imageChange');
+                    if (file.value) {
+                        data['imageId'] = file.value || '';
+                        data['imageName'] = file.name || '';
+
+                    }
+                }
+
             }
-
+            data[name] = value;
+        }
+        formData.append('data', JSON.stringify(data));
 
         if (data.id) {
-                fetch('/notes', {
-                    method: 'PUT',
-                    body: JSON.stringify(data),
-                    headers: {
-                        'Content-Type': 'application/json',
-                    }
-                }).finally(() => {
+            fetch('/notes', {
+                method: 'PUT',
+                body: formData
+            }).then(res => {
+                window.location.href = '/';
+            }).finally(() => {
 
-                    console.log("Go to space and Update!!!!!!");
+                console.log("Go to space and Update!!!!!!");
+            });
 
-                });
-               //window.location.href = '/';
 
-            } else {
-                fetch('/notes', {
-                    method: 'POST',
-                    body: formData,
-                }).then(r => {
-                    console.log(r);
-
-                }).finally(() => {
-                     window.location.href = '/';
-                    console.log("Go to space!!!!!!");
-                });
-
-            }
-       // document.getElementById("img").setAttribute("src", `image/'${data.nameImg}'`);
+        } else {
+            fetch('/notes', {
+                method: 'POST',
+                body: formData,
+            }).then(r => {
+                console.log(r);
+                window.location.href = '/';
+            }).finally(() => {
+                console.log("Go to space!!!!!!");
+            });
 
         }
-    );
+
+
+    }
 }
 
 
@@ -57,39 +66,46 @@ function onBtnDeleteClick(event, id, imageId) {
     if (document.getElementById(id)) {
         document.getElementById(id).remove();
     } else {
-        window.location.href = '/';
+        // window.location.href = '/';
     }
 
-    console.log(imageId);
+    if (id !== null && id !== undefined && id.length > 0) {
+        fetch('/notes', {
+            method: 'DELETE',
+            body: JSON.stringify({id: id}),
+            headers: {
+                'Content-Type': 'application/json',
+            }
+        }).then(res => {
+            // window.location.href = '/';
+        }).finally(() => {
+            console.log("Go to space and delete!!!!!!");
+        });
 
-    fetch('/notes', {
-        method: 'DELETE',
-        body: JSON.stringify({id: id}),
-        headers: {
-            'Content-Type': 'application/json',
-        }
-    }).finally(() => {
-        console.log("Go to space and delete!!!!!!");
-    });
+    }
 
-    fetch('/notes/image', {
-        method: 'DELETE',
-        body: JSON.stringify({imageId: imageId}),
-        headers: {
-            'Content-Type': 'application/json',
-        }
-    }).finally(() => {
-        console.log("Go to space and delete!!!!!!");
-    });
+    if (imageId !== null && imageId !== undefined && imageId.length > 0) {
+        fetch('/notes/image', {
+            method: 'DELETE',
+            body: JSON.stringify({imageId: imageId}),
+            headers: {
+                'Content-Type': 'application/json',
+            }
+        }).finally(() => {
+            console.log("Go to space and delete!!!!!!");
+        });
 
-
+    }
 }
 
 
-function onBtnChangeClick(e, id) {
+function onBtnChangeClick(e) {
+    e.preventDefault();
+    let changeImg = document.getElementById('fileChange');
+    changeImg.click();
+}
 
-   // e.preventDefault();
-
-   let img = document.getElementById('fileChange');
-    return img.click()
+function onFileChange(e, id, imageId) {
+    onBtnDeleteClick(e, id, imageId);
+    onFormSubmit(e);
 }
